+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - ?
+## [1.0.0] - 2023-11-25
 
-- 🚧 Work in Progress
+- add hooks for blood moon start/end
+- add hooks for player died
+- add hooks for player enter/join multiplayer
+- add hooks for player new/load game for local/p2p